SPT-0006 - Review derive.py and remove / replace specific functions (#7)
* refactor derive to retain generic functions

* update imports and calls for removed functions

* update unit tests

* fix unit test errors

* reran regression tests
