+++ cishouseholds/pipeline/input_file_stages.py
@@ -9,10 +9,6 @@ from cishouseholds.pipeline.timestamp_map import test_survey_response_data_versi
 from cishouseholds.pipeline.timestamp_map import test_swab_sample_results_datetime_map
 from cishouseholds.pipeline.translate import translate_welsh_survey_responses
 from cishouseholds.pipeline.validation_schema import validation_schemas
-from cishouseholds.pipeline.version_specific_processing.multi_version_preprocessing import assign_has_been_columns
-from cishouseholds.pipeline.version_specific_processing.multi_version_preprocessing import (
-    derive_additional_v1_2_columns,
-)
 from cishouseholds.pipeline.version_specific_processing.test_participant_data_transformations import (
     transform_participant_extract_digital,
 )
@@ -72,9 +68,7 @@ test_survey_response_data_version_2_parameters = {
     "datetime_column_map": test_survey_response_data_version_2_datetime_map,
     "transformation_functions": [
         clean_survey_responses_version_2,
-        derive_additional_v1_2_columns,
         transform_survey_responses_version_2_delta,
-        assign_has_been_columns,
     ],
     "sep": "|",
     "cast_to_double_list": test_survey_response_data_cast_to_double,
