+++ cishouseholds/pipeline/lookup_and_regex_transformations.py
@@ -9,35 +9,6 @@ from cishouseholds.derive import assign_column_to_date_string
 from cishouseholds.derive import assign_column_value_from_multiple_column_map
 from cishouseholds.derive import assign_regex_from_map
 from cishouseholds.derive import assign_regex_match_result
-from cishouseholds.derive import flag_records_for_childcare_v0_rules
-from cishouseholds.derive import flag_records_for_childcare_v1_rules
-from cishouseholds.derive import flag_records_for_childcare_v2_b_rules
-from cishouseholds.derive import flag_records_for_college_v0_rules
-from cishouseholds.derive import flag_records_for_college_v1_rules
-from cishouseholds.derive import flag_records_for_college_v2_rules
-from cishouseholds.derive import flag_records_for_furlough_rules_v0
-from cishouseholds.derive import flag_records_for_furlough_rules_v1_a
-from cishouseholds.derive import flag_records_for_furlough_rules_v1_b
-from cishouseholds.derive import flag_records_for_furlough_rules_v2_a
-from cishouseholds.derive import flag_records_for_furlough_rules_v2_b
-from cishouseholds.derive import flag_records_for_not_working_rules_v0
-from cishouseholds.derive import flag_records_for_not_working_rules_v1_a
-from cishouseholds.derive import flag_records_for_not_working_rules_v1_b
-from cishouseholds.derive import flag_records_for_not_working_rules_v2_a
-from cishouseholds.derive import flag_records_for_not_working_rules_v2_b
-from cishouseholds.derive import flag_records_for_retired_rules
-from cishouseholds.derive import flag_records_for_school_v2_rules
-from cishouseholds.derive import flag_records_for_self_employed_rules_v0
-from cishouseholds.derive import flag_records_for_self_employed_rules_v1_a
-from cishouseholds.derive import flag_records_for_self_employed_rules_v1_b
-from cishouseholds.derive import flag_records_for_self_employed_rules_v2_a
-from cishouseholds.derive import flag_records_for_self_employed_rules_v2_b
-from cishouseholds.derive import flag_records_for_uni_v0_rules
-from cishouseholds.derive import flag_records_for_uni_v1_rules
-from cishouseholds.derive import flag_records_for_uni_v2_rules
-from cishouseholds.derive import flag_records_for_work_from_home_rules
-from cishouseholds.derive import flag_records_for_work_location_null
-from cishouseholds.derive import flag_records_for_work_location_student
 from cishouseholds.derive import regex_match_result
 from cishouseholds.edit import rename_column_names
 from cishouseholds.expressions import array_contains_any
@@ -52,6 +23,35 @@ from cishouseholds.regex.healthcare_regex import patient_facing_pattern
 from cishouseholds.regex.healthcare_regex import priority_map
 from cishouseholds.regex.healthcare_regex import roles_map
 from cishouseholds.regex.healthcare_regex import social_care_classification
+from cishouseholds.regex.regex_flags import flag_records_for_childcare_v0_rules
+from cishouseholds.regex.regex_flags import flag_records_for_childcare_v1_rules
+from cishouseholds.regex.regex_flags import flag_records_for_childcare_v2_b_rules
+from cishouseholds.regex.regex_flags import flag_records_for_college_v0_rules
+from cishouseholds.regex.regex_flags import flag_records_for_college_v1_rules
+from cishouseholds.regex.regex_flags import flag_records_for_college_v2_rules
+from cishouseholds.regex.regex_flags import flag_records_for_furlough_rules_v0
+from cishouseholds.regex.regex_flags import flag_records_for_furlough_rules_v1_a
+from cishouseholds.regex.regex_flags import flag_records_for_furlough_rules_v1_b
+from cishouseholds.regex.regex_flags import flag_records_for_furlough_rules_v2_a
+from cishouseholds.regex.regex_flags import flag_records_for_furlough_rules_v2_b
+from cishouseholds.regex.regex_flags import flag_records_for_not_working_rules_v0
+from cishouseholds.regex.regex_flags import flag_records_for_not_working_rules_v1_a
+from cishouseholds.regex.regex_flags import flag_records_for_not_working_rules_v1_b
+from cishouseholds.regex.regex_flags import flag_records_for_not_working_rules_v2_a
+from cishouseholds.regex.regex_flags import flag_records_for_not_working_rules_v2_b
+from cishouseholds.regex.regex_flags import flag_records_for_retired_rules
+from cishouseholds.regex.regex_flags import flag_records_for_school_v2_rules
+from cishouseholds.regex.regex_flags import flag_records_for_self_employed_rules_v0
+from cishouseholds.regex.regex_flags import flag_records_for_self_employed_rules_v1_a
+from cishouseholds.regex.regex_flags import flag_records_for_self_employed_rules_v1_b
+from cishouseholds.regex.regex_flags import flag_records_for_self_employed_rules_v2_a
+from cishouseholds.regex.regex_flags import flag_records_for_self_employed_rules_v2_b
+from cishouseholds.regex.regex_flags import flag_records_for_uni_v0_rules
+from cishouseholds.regex.regex_flags import flag_records_for_uni_v1_rules
+from cishouseholds.regex.regex_flags import flag_records_for_uni_v2_rules
+from cishouseholds.regex.regex_flags import flag_records_for_work_from_home_rules
+from cishouseholds.regex.regex_flags import flag_records_for_work_location_null
+from cishouseholds.regex.regex_flags import flag_records_for_work_location_student
 from cishouseholds.regex.regex_patterns import at_school_pattern
 from cishouseholds.regex.regex_patterns import at_university_pattern
 from cishouseholds.regex.regex_patterns import childcare_pattern
