+++ tests/derive/test_assign_visit_order_from_digital.py
@@ -1,10 +1,10 @@
 import pyspark.sql.functions as F
 from chispa import assert_df_equality
 
-from cishouseholds.derive import assign_visit_order
+from cishouseholds.derive import assign_incremental_order
 
 
-def test_assign_visit_order(spark_session):
+def test_assign_incremental_order(spark_session):
     expected_df = spark_session.createDataFrame(
         data=[
             # fmt: off
@@ -21,10 +21,10 @@ def test_assign_visit_order(spark_session):
     )
     expected_df = expected_df.withColumn("date", F.to_timestamp("date"))
 
-    output_df = assign_visit_order(
+    output_df = assign_incremental_order(
         df=expected_df.drop("count_value"),
         column_name_to_assign="count_value",
-        id="patient_id",
+        id_column="patient_id",
         order_list=["date", "visit_id"],
     )
     assert_df_equality(output_df, expected_df, ignore_nullable=True, ignore_column_order=True, ignore_row_order=True)
