+++ tests/derive/test_flag_records_for_student_rules.py
@@ -1,6 +1,6 @@
 from chispa import assert_df_equality
 
-from cishouseholds.derive import flag_records_for_school_v2_rules
+from cishouseholds.regex.regex_flags import flag_records_for_school_v2_rules
 
 
 def test_flag_records_for_school_v2_rules(spark_session):
