+++ tests/derive/test_flag_records_for_university_rules.py
@@ -1,7 +1,7 @@
 from chispa import assert_df_equality
 
-from cishouseholds.derive import flag_records_for_uni_v0_rules
-from cishouseholds.derive import flag_records_for_uni_v2_rules
+from cishouseholds.regex.regex_flags import flag_records_for_uni_v0_rules
+from cishouseholds.regex.regex_flags import flag_records_for_uni_v2_rules
 
 
 def test_flag_records_for_uni_v2_rules(spark_session):
