+++ tests/pipeline/test_survey_response_data_version_1/processed_responses_phm_schema_regression.yml
@@ -320,6 +320,14 @@ fields:
   nullable: true
   type: timestamp
 - metadata: {}
+  name: phm_think_had_respiratory_infection
+  nullable: true
+  type: string
+- metadata: {}
+  name: phm_think_had_respiratory_infection_type
+  nullable: true
+  type: string
+- metadata: {}
   name: think_had_covid_onset_date
   nullable: true
   type: timestamp
@@ -339,6 +347,10 @@ fields:
   name: think_had_covid_no_symptoms_list_2
   nullable: true
   type: string
+- metadata: {}
+  name: phm_think_had_flu_onset_date
+  nullable: true
+  type: timestamp
 - metadata: {}
   name: think_had_flu_any_symptom_list_1
   nullable: true
@@ -740,95 +752,7 @@ fields:
   nullable: true
   type: string
 - metadata: {}
-  name: phm_think_had_other_infection
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_runny_nose_or_sneezing
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_loss_of_smell
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_loss_of_taste
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_sore_throat
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_cough
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_shortness_of_breath
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_noisy_breathing
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_abdominal_pain
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_nausea_or_vomiting
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_diarrhoea
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_loss_of_appetite
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_headache
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_muscle_ache
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_fatigue
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_fever
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_more_trouble_sleeping
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_memory_loss_or_confusion
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_difficulty_concentrating
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_anxiety
-  nullable: true
-  type: string
-- metadata: {}
-  name: think_have_new_or_worse_symptom_low_mood
-  nullable: true
-  type: string
-- metadata: {}
-  name: survey_completion_status
-  nullable: true
-  type: string
-- metadata: {}
-  name: phm_think_had_flu_onset_date
+  name: phm_think_had_covid
   nullable: true
   type: string
 - metadata: {}
@@ -836,15 +760,7 @@ fields:
   nullable: true
   type: string
 - metadata: {}
-  name: phm_think_had_respiratory_infection
-  nullable: true
-  type: string
-- metadata: {}
-  name: phm_think_had_respiratory_infection_type
-  nullable: true
-  type: string
-- metadata: {}
-  name: phm_think_had_covid
+  name: phm_think_had_other_infection
   nullable: true
   type: string
 - metadata: {}
@@ -929,6 +845,86 @@ fields:
   type: string
 - metadata: {}
   name: think_have_covid_symptom_low_mood
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_runny_nose_or_sneezing
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_loss_of_smell
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_loss_of_taste
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_sore_throat
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_cough
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_shortness_of_breath
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_noisy_breathing
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_abdominal_pain
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_nausea_or_vomiting
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_diarrhoea
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_loss_of_appetite
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_headache
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_muscle_ache
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_fatigue
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_fever
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_more_trouble_sleeping
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_memory_loss_or_confusion
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_difficulty_concentrating
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_anxiety
+  nullable: true
+  type: string
+- metadata: {}
+  name: think_have_new_or_worse_symptom_low_mood
   nullable: true
   type: string
 - metadata: {}
@@ -1357,10 +1353,6 @@ fields:
   type: integer
 - metadata: {}
   name: cis_covid_vaccine_number_of_doses
-  nullable: true
-  type: string
-- metadata: {}
-  name: participant_completion_window_status
   nullable: true
   type: string
 type: struct
