+++ tests/pipeline/test_survey_response_data_version_2/processed_responses_v2_schema_regression.yml
@@ -1068,14 +1068,6 @@ fields:
   nullable: true
   type: string
 - metadata: {}
-  name: work_social_care
-  nullable: true
-  type: string
-- metadata: {}
-  name: work_health_care_patient_facing
-  nullable: true
-  type: string
-- metadata: {}
   name: cis_covid_vaccine_number_of_doses_raw
   nullable: true
   type: string
@@ -1102,13 +1094,5 @@ fields:
 - metadata: {}
   name: work_status_v1
   nullable: true
-  type: string
-- metadata: {}
-  name: household_been_care_home_last_28_days
-  nullable: false
-  type: string
-- metadata: {}
-  name: household_been_hospital_last_28_days
-  nullable: false
   type: string
 type: struct
